@@ -6,6 +6,9 @@
         "ext-ctype": "*",
         "ext-iconv": "*",
         "doctrine/annotations": "^1.11",
+        "doctrine/doctrine-bundle": "^2.3",
+        "doctrine/doctrine-migrations-bundle": "^3.2",
+        "doctrine/orm": "^2.8",
         "phpdocumentor/reflection-docblock": "^5.2",
         "symfony/console": "5.1.*",
         "symfony/dotenv": "5.1.*",
@@ -21,7 +24,11 @@
         "symfony/yaml": "5.1.*"
     },
     "require-dev": {
-        "symfony/maker-bundle": "^1.23"
+        "phpunit/phpunit": "^9.5",
+        "symfony/browser-kit": "5.1.*",
+        "symfony/css-selector": "5.1.*",
+        "symfony/maker-bundle": "^1.23",
+        "symfony/phpunit-bridge": "^7.0"
     },
     "config": {
         "optimize-autoloader": true,
