@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Service;
+namespace App\Service\Sender;
 
 use App\Model\Message;
 
 interface SenderInterface
 {
-    public function supports(Message $message);
+    public function supports(Message $message): bool;
     public function send(Message $message);
 }
